* feat: rough implementation of service, that updates the players actio…

* chore: converted ActionControlHandler (former PlayerControlService) t…

* Merge remote-tracking branch 'origin/main' into 45-prevent-player-fro…

* chore: solved merge conflicts

* chore: calculate animation duration in AnimationClip

* Merge remote-tracking branch 'origin/main' into 45-prevent-player-fro…

* chore: convert PlayerInputSystem into three separate systems

+++ unnamed/src/Components/Rendering/AlignedCharacter.cs
@@ -5,6 +5,6 @@ namespace unnamed.Components.Rendering;
 public struct AlignedCharacter
 {
     public CharacterDirection CharacterDirection;
-    public int ActionIndex;
+    public byte ActionIndex;
     public CharacterType CharacterType;
 }
